Create now checks to make sure the file name does not already exist. Read will correctly read the selected filename now.

+++ source/os/fileSystemDeviceDriver.ts
@@ -35,25 +35,28 @@ module TSOS {
         //Will create a file
         public create(fileName:string):void{
             var done = false;
-            //Loop thorough an find free space
-            for(var x = 0; x < this.tracks; x++) {
+            //Loop thorough an find free space from 000 - 077
+            for(var x = 0; x <= 0; x++) {
                 for (var y = 0; y < this.sectors; y++) {
                     for (var z = 0; z < this.blocks; z++) {
                         var key = x + "" + y + "" + z;
-                        var meta = sessionStorage.getItem(key).substr(0, 1);
-                        if (meta == "0") {
-                            var data = "1---" + this.stringToHex(fileName);
-                            while(data.length < 64){
-                                data = data + "0";
+                        if(key != "000") {
+                            var meta = sessionStorage.getItem(key).substr(0, 1);
+                            if (meta == "0" && this.doesFileExist(fileName) == false) {
+                                var data = "1---" + this.stringToHex(fileName);
+                                while (data.length < 64) {
+                                    data = data + "0";
+                                }
+                                //Place the file on the disk
+                                sessionStorage.setItem(key, data);
+                                _FileSystemDisplay.updateDisplay();
+                                _success = true;
+                                //If the file was created break out of the loop
+                                done = true;
+                                break;
                             }
-                            //Place the file on the disk
-                            sessionStorage.setItem(key, data);
-                            _FileSystemDisplay.updateDisplay();
-                            _success = true;
-                            //If the file was created break out of the loop
-                            done = true;
-                            break;
                         }
+
                     }
                     //If the file was created break out of the loop
                     if(done){
@@ -69,19 +72,25 @@ module TSOS {
 
         public read(fileName:string):void{
             var done = false;
-            //Loop thorough an find free space
+            //Loop through and find the file
             for(var x = 0; x < this.tracks; x++) {
                 for (var y = 0; y < this.sectors; y++) {
                     for (var z = 0; z < this.blocks; z++) {
                         var key = x + "" + y + "" + z;
                         var meta = sessionStorage.getItem(key).substr(0, 1);
+                        //Make sure meta is in use
                         if (meta == "1") {
-                            _Console.putText(""+this.hexToString(sessionStorage.getItem(key)) + " ");
-                            _Console.advanceLine();
-                            _success = true;
-                            //If the file was created break out of the loop
-                            done = true;
-                            break;
+                            var compare = this.getFileName(key);
+                            var fileName1 = ""+fileName;
+                            //Make sure file names match
+                            if(compare === fileName1) {
+                                _Console.putText("" + this.hexToString(sessionStorage.getItem(key)) + " ");
+                                _Console.advanceLine();
+                                _success = true;
+                                //If the file was created break out of the loop
+                                done = true;
+                                break;
+                            }
                         }
                     }
                     //If the file was created break out of the loop
@@ -97,7 +106,57 @@ module TSOS {
         }
 
         public delete(fileName:string):void{
+            
+        }
 
+        //Will let us know if a file name is already in use
+        public doesFileExist(fileName:string){
+            var done = false;
+            //Loop through and find the file in 000 - 077
+            for(var x = 0; x <=0; x++) {
+                for (var y = 0; y < this.sectors; y++) {
+                    for (var z = 0; z < this.blocks; z++) {
+                        var key = x + "" + y + "" + z;
+                        if(key != "000") {
+                            //Make sure meta is in use
+                            var meta = sessionStorage.getItem(key).substr(0, 1);
+                            if(meta == "1"){
+                                var compare = this.getFileName(key);
+                                var fileName1 = ""+fileName;
+                                //Make sure file names match
+                                if(compare === fileName1) {
+                                    return true;
+                                    done = true;
+                                    break;
+                                }
+                            }
+                        }
+                    }
+                    //If the file was created break out of the loop
+                    if(done){
+                        break;
+                    }
+                }
+                //If the file was created break out of the loop
+                if(done){
+                    break;
+                }
+            }
+            if(!done){
+                return false;
+            }
+        }
+
+        //Will get the file name from the key provided
+        public getFileName(key){
+            //Get the file name.
+            var fileName = sessionStorage.getItem(key);
+            fileName = fileName.replace(/0+$/,"");
+            if(fileName.length % 2){
+                fileName = fileName + 0;
+            }
+            //return the file name
+            return this.hexToString(fileName);
         }
 
         //Used to put a string into hex
@@ -112,19 +171,19 @@ module TSOS {
             return newString;
         }
 
+        //Will take hex and make it a string
         public hexToString(hexStr : string){
             var newString = "";
             var hexStr1 = ""+hexStr;
             //Loop through the hex and change it to string
             for(var x = 4; hexStr1.length > x; x = x+2){
                 if(hexStr1.length > x+1) {
-
                     var piece1 = String.fromCharCode(_MemoryManager.hexToDec(hexStr1.substr(x, 2)));
                     newString = newString + piece1;
                 }
             }
+            //return the string
             return newString;
-
         }
